Update manticore challenge
