+++ .travis.yml
@@ -11,6 +11,8 @@ python:
 install:
   - pip install manticore[native]
   - pip install nose
+  - pip install r2pipe
+  - apt install radare2
 
 script:
   - nosetests
