+++ test_manticore_challenge/manticore_challenge.c
